# Conflicts:
#	bitstring/src/main/java/org/ton/java/bitstring/BitString.java
#	cell/src/main/java/org/ton/java/cell/CellSlice.java
#	cell/src/test/java/org/ton/java/cell/TestCellBuilder.java
#	tonlib/src/test/java/org/ton/java/tonlib/TestTonlibJson.java

+++ cell/src/main/java/org/ton/java/cell/DeserializeCellDataResult.java
@@ -2,9 +2,9 @@ package org.ton.java.cell;
 
 public class DeserializeCellDataResult {
     Cell cell;
-    byte[] cellsData;
+    int[] cellsData;
 
-    DeserializeCellDataResult(Cell cell, byte[] cellsData) {
+    DeserializeCellDataResult(Cell cell, int[] cellsData) {
         this.cell = cell;
         this.cellsData = cellsData;
     }
